healthcare-rag-backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
requests
python-multipart
